+++ src/collection.ts
@@ -1,8 +1,9 @@
 import { v1 } from "uuid";
 
 import * as Structs from "./structs";
+import * as Entity from "./entity";
 
-export function createCollection(
+export function createNew(
   Label: Structs.Collection["Label"] = null,
   Entities?: Structs.Collection["Entities"],
   Relations?: Structs.Collection["Relations"],
@@ -14,6 +15,16 @@ export function createCollection(
     Entities: Entities === undefined ? {} : Entities,
     Relations: Relations === undefined ? [] : Relations,
   };
+
+  if (Relations === undefined) {
+    for (const entityID in collection.Entities) {
+      const uniqueRelations = Entity.getUniqueRelations(
+        collection.Entities[entityID],
+        collection.Relations
+      );
+      collection.Relations.push(...uniqueRelations);
+    }
+  }
 
   return collection;
 }
