+++ src/driver.ts
@@ -1,3 +1,7 @@
 /**
  * database driver: manages all entities loaded to memory.
  */
+
+/**
+ * todo: when a relation is used to claim a relation on a entity driver should ensure that that relation exists in collection.relations array for any collection that entity belongs to.
+ */
