+++ src/query.ts
@@ -14,7 +14,7 @@ export class Query {
     this.Collection = Collection;
   }
   hasLabel(Label: Structs.Entity["Label"]): Query {
-    const newCollection: Structs.Collection = Collection.createCollection(
+    const newCollection: Structs.Collection = Collection.createNew(
       this.Collection.Label,
       undefined,
       undefined,
